deploy/requirements.txt
gunicorn==20.1.0
uvicorn==0.15.0
fastapi==0.70.0
pandas==1.0.5
numpy==1.18.5
scikit-learn==0.23.1
jieba==0.42.1
pytest==5.4.3
pytest-html==3.1.1
torchaudio
onnxruntime-gpu
